add thunk and thunkaction

+++ routes/api/questions.js
@@ -2,6 +2,7 @@ const express = require("express")
 const router = express.Router();
 const passport = require("passport");
 const validateQuestionInput = require('../../validation/question');
+const validateResponse = require('../../validation/response')
 const Question = require('../../models/Question');
 const User = require('../../models/User');
 
@@ -11,7 +12,6 @@ router.get('/test', (req, res) => {
     res.json({ msg: "This is the question route" })
 })
 //test
-
 
 //retreiving all the questions
 router.get('/',(req,res) => {
@@ -26,8 +26,9 @@ router.get('/',(req,res) => {
 router.get('/:id',(req,res)=>{
     Question.findById(req.params.id)
     .then(question => res.json(question))
-    .catch(err => res.status(404).json(err))
+    .catch(err => res.status(404).json("question not found"))
 })
+
 
 //posting questions
 router.post('/',passport.authenticate('jwt',{session:false}),(req,res) =>{
@@ -99,7 +100,7 @@ router.patch("/:id", passport.authenticate('jwt',{session:false}), async (req, r
 
         }
 	} catch(err) {
-        res.json(err)
+       
         res.status(404).json({
             error: "Question doesn't exist!"
         })
@@ -107,24 +108,71 @@ router.patch("/:id", passport.authenticate('jwt',{session:false}), async (req, r
     
 })
 
-
-
-
 //deleting a question
 router.delete("/:id", passport.authenticate('jwt',{session:false}), async (req, res) => {
 
-    // console.log('1', req.user.id)
     const question = await Question.findOne({ _id: req.params.id })
-    // console.log('2', question.user)
 
-    if (`${question.user}` === req.user.id){
-        Question.findByIdAndDelete(req.params.id)
-        .then(() => res.json("Question deleted"))
-        .catch(err => res.status(404).json(err))
-    } else{
-        res.status(404).json({error: 'Incorrect user'})
+    if(question) {
+        if (`${question.user}` === req.user.id){
+            Question.findByIdAndDelete(req.params.id)
+            .then(() => res.json(question))
+            .catch(err => res.status(404).json(err))
+        } else{
+            res.status(404).json({error: 'Incorrect user'})
+        }
+    } else {
+        res.json("question not found")
     }
-    
+})
+
+
+// responses
+router.post("/:id/responses", passport.authenticate('jwt',{session:false}), async (req, res) => {
+    let question = await Question.findById(req.params.id)
+
+    const { errors, isValid } = validateResponse(req.body);
+
+    if(question) {
+
+        if(!isValid) {
+
+            return res.status(400).json(errors)
+
+        } else {
+
+            question.responses.push(Object.assign(req.body, {user: req.user.id}))
+            question.save( function (err) {
+                if (!err) res.json(question)
+            })
+
+            // push ref into users document 
+        }
+
+    } else {
+        res.json("question does not exist.")
+    }
+})
+
+router.delete("/:questionId/responses/:responseId", passport.authenticate('jwt',{session:false}), async (req, res) => {
+    let question = await Question.findById(req.params.questionId);
+    let response = await question.responses.id(req.params.responseId)
+
+    if(question && response) {
+
+        if (`${response.user}` === req.user.id){
+            console.log(req.params.responseId)
+            question.responses.id(req.params.responseId).remove();
+            question.save(function (err) {
+                res.json(response)
+            })
+        } else{
+            res.status(404).json('You can only delete your own responses.')
+        }
+        
+    } else {
+        res.json("question and/or response does not exist.")
+    }
 })
 
 module.exports = router;
